requirements.txt
Django==1.8.2
MySQL-python==1.2.5
Pygments==2.0.2
beautifulsoup4==4.4.0
django-angular==0.7.15
djangorestframework==3.1.3
djangular==0.2.7
selenium==2.46.0
wsgiref==0.1.2
